+++ changelog.md
@@ -1,4 +1,4 @@
 ## Changes for 20231021.0.1
 
-* testing compatibility with nvda-2024.1.0;
-* Updated translations.
+* Added support for the speak on demand feature introduced with NVDA-2024.1;
+* Fixed some bugs that appeared when consulting the start and end of the selection as well as the actual, remaining and total duration.